# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API HỆ THỐNG CỬA HÀNG ĐIỆN THOẠI PHÚC MOBILE",
        "description": "Hệ thống API quản lý cửa hàng kinh doanh điện thoại di động và phụ kiện công nghệ PHÚC MOBILE. Hỗ trợ đầy đủ phân quyền người dùng (RBAC) và chuẩn hóa dữ liệu quan hệ 3NF.\n *     \n *     ### Danh sách Vai trò (Roles):\n *     - **superadmin**: Quyền cao nhất, quản lý toàn bộ hệ thống bao gồm cả nhân viên và khách hàng.\n *     - **admin**: Quản lý nghiệp vụ chính (Điện thoại, Hãng sản xuất, Danh mục, Mã giảm giá, Đơn hàng, Đánh giá).\n *     - **staff**: Nhân viên vận hành cửa hàng, xử lý trạng thái đơn hàng và phê duyệt đánh giá sản phẩm.\n *     - **user**: Khách hàng (Xem sản phẩm, thêm giỏ hàng, đặt mua điện thoại và gửi đánh giá).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server (Tự động nhận diện Host)"
        }
    ],
    "paths": {
...
        "/cart": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "summary": "Lấy thông tin giỏ hàng",
                "operationId": "294e093bf29c111c01d8402121fb4f43",
                "parameters": [
                    {
                        "name": "X-Session-ID",
                        "in": "header",
                        "description": "Session ID cho khách vãng lai",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lấy thông tin thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "cart_id": {
                                                    "type": "integer"
                                                },
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "total_quantity": {
                                                    "type": "integer"
                                                },
                                                "total_price": {
                                                    "type": "number"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/cart/add": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Thêm sản phẩm vào giỏ hàng",
                "operationId": "567ccee74d03c3c5737e69a6cae4a3d5",
                "parameters": [
                    {
                        "name": "X-Session-ID",
                        "in": "header",
                        "description": "Session ID cho khách vãng lai",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id",
                                    "quantity"
                                ],
                                "properties": {
                                    "product_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "product_variant_id": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Thêm sản phẩm thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Đã thêm sản phẩm vào giỏ hàng."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dữ liệu không hợp lệ / Hết hàng"
                    },
                    "500": {
                        "description": "Lỗi server"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/cart/update/{itemId}": {
            "put": {
                "tags": [
                    "Cart"
                ],
                "summary": "Cập nhật số lượng sản phẩm trong giỏ",
                "operationId": "4c7bade15314466de651458aa458826b",
                "parameters": [
                    {
                        "name": "itemId",
                        "in": "path",
                        "description": "ID của chi tiết giỏ hàng (CartItem ID)",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "X-Session-ID",
                        "in": "header",
                        "description": "Session ID cho khách vãng lai",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "quantity"
                                ],
                                "properties": {
                                    "quantity": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Cập nhật số lượng thành công."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Không đủ số lượng tồn kho"
                    },
                    "404": {
                        "description": "Không tìm thấy giỏ hàng hoặc sản phẩm"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/cart/remove/{itemId}": {
            "delete": {
                "tags": [
                    "Cart"
                ],
                "summary": "Xóa 1 sản phẩm khỏi giỏ hàng",
                "operationId": "ef0a4c2072529ccae936ff74a02ab873",
                "parameters": [
                    {
                        "name": "itemId",
                        "in": "path",
                        "description": "ID của chi tiết giỏ hàng (CartItem ID)",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "X-Session-ID",
                        "in": "header",
                        "description": "Session ID cho khách vãng lai",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xóa thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Đã xóa sản phẩm khỏi giỏ hàng."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy giỏ hàng hoặc sản phẩm"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/cart/clear": {
            "delete": {
                "tags": [
                    "Cart"
                ],
                "summary": "Xóa toàn bộ giỏ hàng",
                "operationId": "6c16eeec461bf91618968113b533465a",
                "parameters": [
                    {
                        "name": "X-Session-ID",
                        "in": "header",
                        "description": "Session ID cho khách vãng lai",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xóa thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Đã xóa tất cả sản phẩm trong giỏ hàng."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Nhập JWT token vào đây để thực hiện các yêu cầu có bảo mật.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}